requirements.txt
fire>=0.1.3
regex==2017.4.5
requests==2.21.0
tqdm==4.31.1
tensorflow==1.14.0
discord
pillow